+++ packages/twenty-front/src/modules/object-record/record-show/components/RecordShowContainer.tsx
@@ -1,47 +1,31 @@
 import groupBy from 'lodash.groupby';
-import { useRecoilState, useRecoilValue } from 'recoil';
 
 import { ActivityTargetsInlineCell } from '@/activities/inline-cell/components/ActivityTargetsInlineCell';
 import { Note } from '@/activities/types/Note';
 import { Task } from '@/activities/types/Task';
 import { InformationBannerDeletedRecord } from '@/information-banner/components/deleted-record/InformationBannerDeletedRecord';
 import { useGetStandardObjectIcon } from '@/object-metadata/hooks/useGetStandardObjectIcon';
-import { useLabelIdentifierFieldMetadataItem } from '@/object-metadata/hooks/useLabelIdentifierFieldMetadataItem';
-import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { useObjectMetadataItems } from '@/object-metadata/hooks/useObjectMetadataItems';
 import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
 import { formatFieldMetadataItemAsColumnDefinition } from '@/object-metadata/utils/formatFieldMetadataItemAsColumnDefinition';
-import { useUpdateOneRecord } from '@/object-record/hooks/useUpdateOneRecord';
-import {
-  FieldContext,
-  RecordUpdateHook,
-  RecordUpdateHookParams,
-} from '@/object-record/record-field/contexts/FieldContext';
+import { FieldContext } from '@/object-record/record-field/contexts/FieldContext';
 import { RecordInlineCell } from '@/object-record/record-inline-cell/components/RecordInlineCell';
 import { PropertyBox } from '@/object-record/record-inline-cell/property-box/components/PropertyBox';
 import { PropertyBoxSkeletonLoader } from '@/object-record/record-inline-cell/property-box/components/PropertyBoxSkeletonLoader';
 import { InlineCellHotkeyScope } from '@/object-record/record-inline-cell/types/InlineCellHotkeyScope';
+import { useRecordShowContainerActions } from '@/object-record/record-show/hooks/useRecordShowContainerActions';
+import { useRecordShowContainerData } from '@/object-record/record-show/hooks/useRecordShowContainerData';
 import { RecordDetailDuplicatesSection } from '@/object-record/record-show/record-detail-section/components/RecordDetailDuplicatesSection';
 import { RecordDetailRelationSection } from '@/object-record/record-show/record-detail-section/components/RecordDetailRelationSection';
-import { recordLoadingFamilyState } from '@/object-record/record-store/states/recordLoadingFamilyState';
-import { recordStoreFamilyState } from '@/object-record/record-store/states/recordStoreFamilyState';
-import { recordStoreIdentifierFamilySelector } from '@/object-record/record-store/states/selectors/recordStoreIdentifierSelector';
-import { ObjectRecord } from '@/object-record/types/ObjectRecord';
 import { isFieldCellSupported } from '@/object-record/utils/isFieldCellSupported';
-import { useIsPrefetchLoading } from '@/prefetch/hooks/useIsPrefetchLoading';
 import { ShowPageContainer } from '@/ui/layout/page/ShowPageContainer';
-import { ShowPageLeftContainer } from '@/ui/layout/show-page/components/ShowPageLeftContainer';
-import { ShowPageRightContainer } from '@/ui/layout/show-page/components/ShowPageRightContainer';
 import { ShowPageSummaryCard } from '@/ui/layout/show-page/components/ShowPageSummaryCard';
 import { ShowPageSummaryCardSkeletonLoader } from '@/ui/layout/show-page/components/ShowPageSummaryCardSkeletonLoader';
 import { useIsMobile } from '@/ui/utilities/responsive/hooks/useIsMobile';
-import {
-  FieldMetadataType,
-  FileFolder,
-  useUploadImageMutation,
-} from '~/generated/graphql';
+
+import { useRecordShowContainerTabs } from '@/object-record/record-show/hooks/useRecordShowContainerTabs';
+import { ShowPageSubContainer } from '@/ui/layout/show-page/components/ShowPageSubContainer';
+import { FieldMetadataType } from '~/generated/graphql';
 import { isDefined } from '~/utils/isDefined';
-import { isUndefinedOrNull } from '~/utils/isUndefinedOrNull';
 
 type RecordShowContainerProps = {
   objectNameSingular: string;
@@ -58,70 +42,25 @@ export const RecordShowContainer = ({
   isInRightDrawer = false,
   isNewRightDrawerItemLoading = false,
 }: RecordShowContainerProps) => {
-  const { objectMetadataItem } = useObjectMetadataItem({
+  const {
+    recordFromStore,
+    recordLoading,
+    objectMetadataItem,
+    labelIdentifierFieldMetadataItem,
+    isPrefetchLoading,
+    recordIdentifier,
+    objectMetadataItems,
+  } = useRecordShowContainerData({
     objectNameSingular,
+    objectRecordId,
   });
 
-  const { objectMetadataItems } = useObjectMetadataItems();
-
-  const { labelIdentifierFieldMetadataItem } =
-    useLabelIdentifierFieldMetadataItem({
+  const { onUploadPicture, useUpdateOneObjectRecordMutation } =
+    useRecordShowContainerActions({
       objectNameSingular,
+      objectRecordId,
+      recordFromStore,
     });
-
-  const [recordLoading] = useRecoilState(
-    recordLoadingFamilyState(objectRecordId),
-  );
-
-  const [recordFromStore] = useRecoilState<ObjectRecord | null>(
-    recordStoreFamilyState(objectRecordId),
-  );
-
-  const recordIdentifier = useRecoilValue(
-    recordStoreIdentifierFamilySelector({
-      objectNameSingular,
-      recordId: objectRecordId,
-    }),
-  );
-  const [uploadImage] = useUploadImageMutation();
-  const { updateOneRecord } = useUpdateOneRecord({ objectNameSingular });
-
-  const useUpdateOneObjectRecordMutation: RecordUpdateHook = () => {
-    const updateEntity = ({ variables }: RecordUpdateHookParams) => {
-      updateOneRecord?.({
-        idToUpdate: variables.where.id as string,
-        updateOneRecordInput: variables.updateOneRecordInput,
-      });
-    };
-
-    return [updateEntity, { loading: false }];
-  };
-
-  const onUploadPicture = async (file: File) => {
-    if (objectNameSingular !== 'person') {
-      return;
-    }
-
-    const result = await uploadImage({
-      variables: {
-        file,
-        fileFolder: FileFolder.PersonPicture,
-      },
-    });
-
-    const avatarUrl = result?.data?.uploadImage;
-
-    if (!avatarUrl || isUndefinedOrNull(updateOneRecord) || !recordFromStore) {
-      return;
-    }
-
-    await updateOneRecord({
-      idToUpdate: objectRecordId,
-      updateOneRecordInput: {
-        avatarUrl,
-      },
-    });
-  };
 
   const availableFieldMetadataItems = objectMetadataItem.fields
     .filter(
@@ -163,7 +102,6 @@ export const RecordShowContainer = ({
   const { Icon, IconColor } = useGetStandardObjectIcon(objectNameSingular);
   const isReadOnly = objectMetadataItem.isRemote;
   const isMobile = useIsMobile() || isInRightDrawer;
-  const isPrefetchLoading = useIsPrefetchLoading();
 
   const summaryCard =
     !isNewRightDrawerItemLoading && isDefined(recordFromStore) ? (
@@ -315,6 +253,12 @@ export const RecordShowContainer = ({
     </>
   );
 
+  const tabs = useRecordShowContainerTabs(
+    loading,
+    objectMetadataItem.nameSingular as CoreObjectNameSingular,
+    isInRightDrawer,
+  );
+
   return (
     <>
       {recordFromStore && recordFromStore.deletedAt && (
@@ -324,21 +268,14 @@ export const RecordShowContainer = ({
         />
       )}
       <ShowPageContainer>
-        <ShowPageLeftContainer forceMobile={isMobile}>
-          {!isMobile && summaryCard}
-          {!isMobile && fieldsBox}
-        </ShowPageLeftContainer>
-        <ShowPageRightContainer
+        <ShowPageSubContainer
+          tabs={tabs}
           targetableObject={{
             id: objectRecordId,
             targetObjectNameSingular: objectMetadataItem?.nameSingular,
           }}
-          timeline
-          tasks
-          notes
-          emails
           isInRightDrawer={isInRightDrawer}
-          summaryCard={isMobile ? summaryCard : <></>}
+          summaryCard={summaryCard}
           fieldsBox={fieldsBox}
           loading={isPrefetchLoading || loading || recordLoading}
         />
